Move custom validator to  validation module, use validation constants
