--- src/app/shared/show-required-message.ts
@@ -1,1 +1,0 @@
-export const showRequiredMessage = (label: string) : string => `The ${ label } is required`;
