+++ src/app/shared/validation/constants/validation-messages.constants.ts
@@ -1,0 +1,2 @@
+export const requiredMessage = 'The {{ label }} is required';
+export const passwordMismatchMessage = 'Passwords are different';
