+++ src/app/shared/validation/validators/password-match-validator.ts
@@ -1,14 +1,19 @@
 import { AbstractControl, ValidationErrors, ValidatorFn } from '@angular/forms';
-import { Subscription } from 'rxjs';
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 
 export function passwordMatchValidator(originalPasswordId: string): ValidatorFn {
   return (confirmedPasswordControl: AbstractControl): ValidationErrors | null => {
     const passwordControl = confirmedPasswordControl.root.get(originalPasswordId);
     if (passwordControl) {
-      const subscription: Subscription = passwordControl.valueChanges.subscribe(() => {
-        confirmedPasswordControl.updateValueAndValidity();
-        subscription.unsubscribe();
-      });
+      const unsubscriber = new Subject();
+      passwordControl.valueChanges
+        .pipe(takeUntil(unsubscriber))
+        .subscribe(() => {
+          confirmedPasswordControl.updateValueAndValidity();
+          unsubscriber.next();
+          unsubscriber.complete();
+        });
     }
     const confirmedPassword = confirmedPasswordControl.value as string;
     const password = passwordControl?.value as string;
